animation2

+++ gameItem.cpp
@@ -9,9 +9,9 @@ Item::Item() {
 	memset(colorId, 0, sizeof(colorId));
 }
 
-Item::Item(int assignedShapeId, int assignedColorId) {
-	for (int x = 0; x < 2; x++) {
-		for (int y = 0; y < 2; y++) {
+Item::Item(short assignedShapeId, short assignedColorId) {
+	for (short x = 0; x < 2; x++) {
+		for (short y = 0; y < 2; y++) {
 			shapeId[0][x][y] = assignedShapeId;
 			colorId[0][x][y] = assignedColorId;
 		}
